add analysis.py to the python material
At the end of the python unit we transferred some functions out of the ipython notebook into analysis.py, an independent file that could be run directly from the command line.

analysis.py relies on hardcoded values for filenames and isn't as reusable as we would like it to be. The example file in the code review unit is similarly structured, but it uses sys.argv to read basic arguments supplied to the script. The more robust way of accepting arguments to a file can be found in python's argparse library, which will let you define and gracefully handle flags, keyword arguments, and positional arguments like other unix utilities we learned about in the unix directory.